Fix issue in SSE event handler
